I006D: Edit brightness path
fix the screen go off and never wake.
before, the bl level was 0 after screen of and stay at 0, even when the power key button is pressed.

And, in the same time, with the old path, the max brightnss was 255, now, it's 1024

Use the custom Asus path to fix it